psv: split input/output modules
