+++ test/Icicle/Test/Sea/Psv.hs
@@ -193,5 +193,5 @@ withSystemTempDirectory template action = do
 
 return []
 tests :: IO Bool
-tests = $forAllProperties $ quickCheckWithResult (stdArgs {maxSuccess = 1000, maxSize = 10, maxDiscardRatio = 10000})
+tests = $forAllProperties $ quickCheckWithResult (stdArgs {maxSuccess = 500, maxSize = 10, maxDiscardRatio = 10000})
 -- tests = $forAllProperties $ quickCheckWithResult (stdArgs {maxSuccess = 1000, maxSize = 10, maxDiscardRatio = 10000})
